pyproject.toml
[project]
name = "causal-oracle"
version = "1.0.0"
description = "Causal inference-driven stock prediction system"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "yfinance>=0.2.40",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "statsmodels>=0.14.0",
    "ta>=0.11.0",
    "hmmlearn>=0.3.2",
    "dowhy>=0.11.0",
    "causal-learn>=0.1.3.8",
    "networkx>=3.2.0",
    "joblib>=1.3.0",
    "pandas-datareader>=0.10.0",
    "scikit-learn>=1.4.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["backend"]
